Update README.md
Xcode 11.3.1 is REQUIRED to make this run currently due to memory access errors.  It is possible to compile with a higher Xcode but you will get runtime errors making it useless currently.  This should be stated in the readme for new users compiling.  Also the command line tools must be set to 12.1 or openSSL will fail to compile as well.